+++ tests/test_datatypes.py
@@ -22,6 +22,7 @@
 from smbus2.smbus2 import i2c_smbus_ioctl_data, union_i2c_smbus_data, i2c_msg, i2c_rdwr_ioctl_data  # noqa: F401
 from smbus2.smbus2 import I2C_SMBUS_BLOCK_MAX, I2C_SMBUS_READ, I2C_SMBUS_BYTE_DATA
 import unittest
+import sys
 
 
 class TestDataTypes(unittest.TestCase):
@@ -62,23 +63,28 @@ class TestDataTypes(unittest.TestCase):
         msg = i2c_msg.read(80, 10)
         self.assertEqual(msg.addr, 80)
         self.assertEqual(msg.len, 10)
+        self.assertEqual(msg.len, len(msg))
         self.assertEqual(msg.flags, 1)
 
     def test_i2c_msg_write(self):
         # Create from list
-        buf = [65, 66, 67, 68]
+        buf = [65, 66, 67, 68, 1, 10, 255]
         msg = i2c_msg.write(81, buf)
         self.assertEqual(msg.addr, 81)
-        self.assertEqual(msg.len, 4)
+        self.assertEqual(msg.len, 7)
+        self.assertEqual(msg.len, len(msg))
         self.assertEqual(msg.flags, 0)
         self.assertListEqual(buf, list(msg))
         # Create from str
-        s = "ABCD"
+        s = "ABCD\x01\n\xFF"
         msg2 = i2c_msg.write(81, s)
         self.assertEqual(msg2.addr, msg.addr)
         self.assertEqual(msg2.len, msg.len)
         self.assertEqual(msg2.flags, msg.flags)
         self.assertListEqual(list(msg), list(msg2))
+        self.assertEqual(str(msg2)[0:4], "ABCD")
+        self.assertGreaterEqual(('%r' % msg2).find(r"ABCD\x01\n\xff"), 0)
+
 
     def test_i2c_msg_iter(self):
         buf = [10, 11, 12, 13]